chore: move category to its own feature
