+++ lib/main.dart
@@ -3,19 +3,19 @@ import "package:provider/provider.dart";
 import "package:tabemashou/core/constants/path.dart";
 import "package:tabemashou/core/theme/theme.dart";
 import "package:tabemashou/core/utils/log/logger.dart";
-import "package:tabemashou/data/category/category_local_source.dart";
 import "package:tabemashou/data/checklist_item/checklist_item_local_source.dart";
 import "package:tabemashou/data/review/review_local_source.dart";
-import "package:tabemashou/domain/category/category_repository_impl.dart";
 import "package:tabemashou/domain/checklist_item/checklist_item_repository_impl.dart";
 import "package:tabemashou/domain/review/review_repository_impl.dart";
-import "package:tabemashou/presentation/providers/category_provider.dart";
+import "package:tabemashou/features/category/data/category_local_source.dart";
+import "package:tabemashou/features/category/domain/category_repository_impl.dart";
+import "package:tabemashou/features/category/presentation/providers/category_provider.dart";
+import "package:tabemashou/features/category/presentation/screens/category_setting_screen.dart";
 import "package:tabemashou/presentation/providers/checklist_item_provider.dart";
 import "package:tabemashou/presentation/providers/review_provider.dart";
 import "package:tabemashou/presentation/screens/about_screen.dart";
 import "package:tabemashou/presentation/screens/main_navigator_screen.dart";
 import "package:tabemashou/presentation/screens/reviews/review_form_screen.dart";
-import "package:tabemashou/presentation/screens/settings/category_setting_screen.dart";
 import "package:tabemashou/presentation/screens/settings/settings_screen.dart";
 import "package:tabemashou/presentation/screens/tabs/categories_screen.dart";
 import "package:tabemashou/presentation/screens/tabs/home_checklist_screen.dart";
