+++ lib/presentation/screens/tabs/home_screen.dart
@@ -2,9 +2,9 @@ import "package:flutter/material.dart";
 import "package:provider/provider.dart";
 import "package:tabemashou/core/constants/path.dart";
 import "package:tabemashou/core/type/checklist.dart";
-import "package:tabemashou/presentation/providers/category_provider.dart";
+import "package:tabemashou/features/category/presentation/providers/category_provider.dart";
+import "package:tabemashou/features/category/presentation/widgets/category_carousel.dart";
 import "package:tabemashou/presentation/providers/checklist_item_provider.dart";
-import "package:tabemashou/presentation/widgets/category/category_carousel.dart";
 import "package:tabemashou/presentation/widgets/checklist/checklist_preview.dart";
 import "package:tabemashou/presentation/widgets/common/section_title.dart";
 import "package:tabemashou/presentation/widgets/common/stat_card.dart";
