+++ lib/presentation/widgets/review/review_config_bottom_sheet/category_config_sort_view.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 import "package:provider/provider.dart";
 import "package:tabemashou/core/type/category.dart";
-import "package:tabemashou/presentation/providers/category_provider.dart";
+import "package:tabemashou/features/category/presentation/providers/category_provider.dart";
 
 class CategoryConfigSortView extends StatefulWidget {
   const CategoryConfigSortView({super.key});
